[Fish S2-Pro] Load fast audio decoder directly
