+++ sglang_omni/models/fishaudio_s2_pro/fish_speech/models/text2semantic/__init__.py
@@ -1,19 +1,12 @@
-"""
-FishQwen3 models with HuggingFace-style interface.
-"""
+"""FishQwen3 serving configuration and inference-only audio decoder."""
 
+from sglang_omni.models.fishaudio_s2_pro.fish_speech.models.text2semantic.audio_decoder import (
+    FishQwen3AudioDecoder,
+)
 from sglang_omni.models.fishaudio_s2_pro.fish_speech.models.text2semantic.configuration import (
     FishQwen3AudioDecoderConfig,
     FishQwen3Config,
     FishQwen3OmniConfig,
-)
-from sglang_omni.models.fishaudio_s2_pro.fish_speech.models.text2semantic.modeling import (
-    FishQwen3AudioDecoder,
-    FishQwen3ForCausalLM,
-    FishQwen3Model,
-    FishQwen3OmniForCausalLM,
-    FishQwen3OmniOutput,
-    FishQwen3PreTrainedModel,
 )
 
 __all__ = [
@@ -21,12 +14,5 @@ __all__ = [
     "FishQwen3Config",
     "FishQwen3AudioDecoderConfig",
     "FishQwen3OmniConfig",
-    # Models
-    "FishQwen3Model",
-    "FishQwen3ForCausalLM",
-    "FishQwen3PreTrainedModel",
-    "FishQwen3OmniForCausalLM",
     "FishQwen3AudioDecoder",
-    # Outputs
-    "FishQwen3OmniOutput",
 ]
